Moved vector into lib
