+++ frontend/src/windows/map_window.ts
@@ -3,7 +3,7 @@ import * as PIXI from "pixi.js";
 import * as ContextMenu from "../lib/contextmenu.ts";
 import { CanvasWindow, registerWindowType } from "./window.ts";
 import { Parameter, GenerateId, LocalPersist } from "../lib/utils.ts";
-import { Vector2 } from "../vector.ts";
+import { Vector2 } from "../lib/vector.ts";
 import { ApiRequest } from "../lib/requests.ts";
 import { MapCanvas } from "../lib/canvas.ts";
 import { ErrorToast } from "../lib/notifications.ts";
